Fix issue with mutiple install the same app for different users
If the same application has been installed for more than one user,
then while uninstallation Smack rule file should remain in the system
until the last "instance" of application is present.

+++ src/common/service_impl.cpp
@@ -391,6 +391,7 @@ int ServiceImpl::appUninstall(const std::string &appId, uid_t uid, bool isSlave)
     std::string smackLabel;
     std::vector<std::string> pkgContents;
     bool appExists = true;
+    bool removeApp = false;
     bool removePkg = false;
     std::string uidstr;
     checkGlobalUser(uid, uidstr);
@@ -420,7 +421,7 @@ int ServiceImpl::appUninstall(const std::string &appId, uid_t uid, bool isSlave)
                 package that the app appears in */
             PrivilegeDb::getInstance().GetAppIdsForPkgId(pkgId, pkgContents);
             PrivilegeDb::getInstance().UpdateAppPrivileges(appId, uid, std::vector<std::string>());
-            PrivilegeDb::getInstance().RemoveApplication(appId, uid, removePkg);
+            PrivilegeDb::getInstance().RemoveApplication(appId, uid, removeApp, removePkg);
 
             if (isSlave) {
                 int ret = MasterReq::CynaraPolicyUpdate(appId, uidstr, std::vector<std::string>());
@@ -462,7 +463,7 @@ int ServiceImpl::appUninstall(const std::string &appId, uid_t uid, bool isSlave)
             if (isSlave) {
                 LogDebug("Delegating Smack rules removal for deleted pkgId " << pkgId <<
                          " to master");
-                int ret = MasterReq::SmackUninstallRules(appId, pkgId, pkgContents, removePkg);
+                int ret = MasterReq::SmackUninstallRules(appId, pkgId, pkgContents, removeApp, removePkg);
                 if (ret != SECURITY_MANAGER_API_SUCCESS) {
                     LogError("Error while processing uninstall request on master: " << ret);
                     return ret;
@@ -472,9 +473,10 @@ int ServiceImpl::appUninstall(const std::string &appId, uid_t uid, bool isSlave)
                     LogDebug("Removing Smack rules for deleted pkgId " << pkgId);
                     SmackRules::uninstallPackageRules(pkgId);
                 }
-
-                LogDebug ("Removing smack rules for deleted appId " << appId);
-                SmackRules::uninstallApplicationRules(appId, pkgId, pkgContents, zoneId);
+                if (removeApp) {
+                    LogDebug("Removing smack rules for deleted appId " << appId);
+                    SmackRules::uninstallApplicationRules(appId, pkgId, pkgContents, zoneId);
+                }
             }
         } catch (const SmackException::Base &e) {
             LogError("Error while removing Smack rules for application: " << e.DumpToString());
